PLANNER-401: Auto-rename org.optaplanner.benchmark.impl.statistic.single -> subsingle
